Now if stdout is not 1, It doesn't print

+++ Userland/SampleCodeModule/include/processManagement.h
@@ -15,7 +15,7 @@
 uint64_t createProcess(char *name, int (*mainF)(int, char **), int argc,
                        char **argv, unsigned int foreground);
 uint64_t createProcessWithFD(char *name, int (*mainF)(int, char **), int argc,
-                       char **argv, unsigned int foreground, uint64_t stdin, uint64_t stdout);
+                       char **argv, unsigned int foreground, unsigned int stdin, unsigned int stdout);
 void waitPID(uint64_t pid);
 
 #endif /* PROCESS_MANAGEMENT_H */
